moved is_safe_path method under generic.filesystem
